Move tests to extracted class's test
